% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkDataResource.R
\docType{package}
\name{dkDataResource}
\alias{dkDataResource}
\alias{dkDataResource-package}
\title{Meteorological package resources.}
\description{
Collecting and transforming meteorological data resources, such as map data, station information, topography data and so on.
}
\details{
dkDataResource
A data package for dkDataResource.

Use \code{data(package='dkDataResource')$results[, 3]} tosee a list of availabledata sets in this data package
    and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{NWSInfo}
}
