% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkTemplate.R
\name{nmcNotebookTemp}
\alias{nmcNotebookTemp}
\title{NMC Clean Markdown Template
Refer to https://github.com/holtzy/epuRate}
\usage{
nmcNotebookTemp(toc = TRUE, code_folding = "show", number_sections = TRUE)
}
\arguments{
\item{toc}{}

\item{code_folding}{}

\item{number_sections}{}
}
\value{

}
\description{
NMC Clean Markdown Template
Refer to https://github.com/holtzy/epuRate
}
