% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkDataResource.R
\docType{data}
\name{NWSInfo}
\alias{NWSInfo}
\title{National weather station information.}
\format{a \code{data.frame} containing the following fields:
\describe{
\item{province}{province name}
\item{ID}{station ID}
\item{name}{station name}
\item{grade}{station grade level}
\item{lat}{latitude}
\item{lon}{longitude}
\item{alt}{altitude}
\item{pressureAlt}{pressure altitude}
}}
\source{
The data comes from the information table of national weather stations (2016).
}
\description{
Detailed description of the data
}
\seealso{
\link{dkDataResource}
}
